update responses for new consensus
