+++ src/types/FoliageBlock.ts
@@ -1,0 +1,9 @@
+export interface FoliageBlock {
+    additions_root: string;
+    filter_hash: string;
+    height: number;
+    prev_block_hash: string;
+    removals_root: string;
+    timestamp: string;
+    transactions_info_hash: string;
+}
