+++ src/types/Proof.ts
@@ -1,0 +1,4 @@
+export interface Proof {
+    witness: string;
+    witness_type: number;
+}
