+++ src/types/RewardClaim.ts
@@ -1,0 +1,5 @@
+export interface RewardClaim {
+    amount: string;
+    parent_coin_info: string;
+    puzzle_hash: string;
+}
