+++ src/types/Vdf.ts
@@ -1,0 +1,10 @@
+export interface VdfOutput {
+    a: string;
+    b: string;
+}
+
+export interface Vdf {
+    challenge: string;
+    number_of_iterations: string;
+    output: VdfOutput;
+}
